# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_request.AdminResetPasswordRequest:
    properties:
      new_password:
        description: 'NewPassword string `json:"new_password" binding:"required,min=8,max=100"`
          // TODO: re-enable min=8 for production'
        maxLength: 100
        minLength: 3
        type: string
    required:
    - new_password
    type: object
...
  github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_request.UserCreateRequest:
    properties:
      address:
        maxLength: 500
        type: string
      bank_account_number:
        maxLength: 100
        type: string
      basic_salary:
        type: number
      birth_date:
        type: string
      birth_place:
        maxLength: 255
        type: string
      department:
        maxLength: 100
        type: string
      email:
        description: 'Email string `json:"email" binding:"required,email,max=255"`
          // TODO: re-enable email validation for production'
        maxLength: 255
        type: string
      employee_code:
        maxLength: 100
        type: string
      employment_status:
        allOf:
        - $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_model.UserEmploymentStatus'
        maxLength: 100
      full_name:
        maxLength: 255
        minLength: 3
        type: string
      gender:
        allOf:
        - $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_model.UserGender'
        enum:
        - male
        - female
        - other
      other_allowance:
        type: number
      phone_number:
        maxLength: 50
        type: string
      position:
        maxLength: 100
        type: string
      position_allowance:
        type: number
      profile_picture_id:
        type: string
      role:
        allOf:
        - $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_model.UserRole'
        enum:
        - admin
        - user
    required:
    - email
    - full_name
    - role
    type: object
  github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_request.UserUpdateProfileRequest:
    properties:
      address:
        maxLength: 500
        type: string
      bank_account_number:
        maxLength: 100
        type: string
      birth_date:
        type: string
      birth_place:
        maxLength: 255
        type: string
      clear_profile_picture:
        type: boolean
      department:
        maxLength: 100
        type: string
      email:
        description: 'Email *string `json:"email" binding:"omitempty,email,max=255"`
          // TODO: re-enable email validation for production'
        maxLength: 255
        type: string
      employee_code:
        maxLength: 100
        type: string
      employment_status:
        allOf:
        - $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_model.UserEmploymentStatus'
        maxLength: 100
      full_name:
        maxLength: 255
        minLength: 3
        type: string
      gender:
        allOf:
        - $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_model.UserGender'
        enum:
        - male
        - female
        - other
      password:
        maxLength: 100
        minLength: 8
        type: string
      phone_number:
        maxLength: 50
        type: string
      position:
        maxLength: 100
        type: string
      profile_picture_id:
        type: string
      role:
        allOf:
        - $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_model.UserRole'
        enum:
        - admin
        - user
    type: object
  github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_request.UserUpdateRequest:
    properties:
      address:
        maxLength: 500
        type: string
      bank_account_number:
        maxLength: 100
        type: string
      basic_salary:
        type: number
      birth_date:
        type: string
      birth_place:
        maxLength: 255
        type: string
      clear_profile_picture:
        type: boolean
      department:
        maxLength: 100
        type: string
      email:
        description: 'Email *string `json:"email" binding:"omitempty,email,max=255"`
          // TODO: re-enable email validation for production'
        maxLength: 255
        type: string
      employee_code:
        maxLength: 100
        type: string
      employment_status:
        allOf:
        - $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_model.UserEmploymentStatus'
        maxLength: 100
      full_name:
        maxLength: 255
        minLength: 3
        type: string
      gender:
        allOf:
        - $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_model.UserGender'
        enum:
        - male
        - female
        - other
      joined_at:
        type: string
      other_allowance:
        type: number
      phone_number:
        maxLength: 50
        type: string
      position:
        maxLength: 100
        type: string
      position_allowance:
        type: number
      profile_picture_id:
        type: string
      role:
        allOf:
        - $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_model.UserRole'
        enum:
        - admin
        - user
    type: object
...
  github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.ActionSuccessResponse:
    properties:
      id:
        type: string
    type: object
...
  github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.ErrorSchema:
    properties:
      code:
        $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.ErrorCode'
      details: {}
      message:
        type: string
    type: object
  github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.ErrorSchema'
      success:
        type: boolean
    type: object
...
  ? github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-array_github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_UserResponse
  : properties:
      data:
        items:
          $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response.UserResponse'
        type: array
      error:
        $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.ErrorSchema'
      success:
        type: boolean
    type: object
...
  ? github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_UserResponse
  : properties:
      data:
        $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response.UserResponse'
      error:
        $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.ErrorSchema'
      success:
        type: boolean
    type: object
  ? github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common_ActionSuccessResponse
  : properties:
      data:
        $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.ActionSuccessResponse'
      error:
        $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.ErrorSchema'
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: afifudin23@example.com
    name: Afifudin
  description: REST API untuk manajemen absensi, karyawan, dan penggajian King Royal.
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: Absensi King Royal API
  version: "1.0"
paths:
...
  /users:
    get:
      description: |-
        Mendapatkan daftar seluruh pengguna yang terdaftar.

        **Query Params (opsional):**
        - `search` (string): Filter berdasarkan nama atau email
        - `role` (string): Filter berdasarkan role — `admin` atau `user`

        **Response 200:** Array data pengguna beserta profil lengkap (posisi, departemen, gaji, dll)
      parameters:
      - description: Cari berdasarkan nama atau email
        in: query
        name: search
        type: string
      - description: Filter berdasarkan role (admin/user)
        in: query
        name: role
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-array_github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_UserResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
      security:
      - BearerAuth: []
      summary: Daftar semua pengguna
      tags:
      - Users
    post:
      consumes:
      - application/json
      description: |-
        Membuat akun pengguna baru beserta data profil karyawan.

        **Request Body:**
        - `full_name` (string, required): Nama lengkap
        - `email` (string, required): Email unik
        - `role` (string, required): Role akun — `admin` atau `user`
        - `employee_code`, `employment_status`, `birth_place`, `birth_date`, `gender` (opsional)
        - `address`, `phone_number`, `position`, `department`, `bank_account_number` (opsional)
        - `basic_salary`, `position_allowance`, `other_allowance` (opsional, angka)

        **Behavior:**
        - Password awal akan di-generate otomatis dan bisa direset oleh admin
        - Mengembalikan daftar semua pengguna setelah user baru berhasil dibuat

        **Response 201:** Array seluruh data pengguna (termasuk yang baru dibuat)
      parameters:
      - description: Data pengguna baru
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_request.UserCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-array_github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
      security:
      - BearerAuth: []
      summary: Buat pengguna baru
      tags:
      - Users
  /users/{user_id}:
    delete:
      description: |-
        Menghapus pengguna berdasarkan ID menggunakan soft delete (data tidak benar-benar terhapus dari database).

        **Path Param:**
        - `user_id` (UUID): ID pengguna yang akan dihapus

        **Response 200:** ID pengguna yang dihapus

        **Error:** 404 jika pengguna tidak ditemukan
      parameters:
      - description: ID pengguna
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common_ActionSuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
      security:
      - BearerAuth: []
      summary: Hapus pengguna
      tags:
      - Users
    get:
      description: |-
        Mendapatkan data lengkap satu pengguna berdasarkan ID.

        **Path Param:**
        - `user_id` (UUID): ID pengguna yang ingin dilihat

        **Response 200:** Data pengguna beserta profil lengkap

        **Error:** 404 jika pengguna tidak ditemukan
      parameters:
      - description: ID pengguna
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
      security:
      - BearerAuth: []
      summary: Detail pengguna
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: |-
        Memperbarui data pengguna tertentu berdasarkan ID. Semua field opsional (partial update).

        **Field yang bisa diupdate:** `full_name`, `email`, `role`, `employee_code`, `employment_status`,
        `birth_place`, `birth_date`, `gender`, `address`, `phone_number`, `position`, `department`,
        `bank_account_number`, `basic_salary`, `position_allowance`, `other_allowance`, `joined_at`,
        `profile_picture_id`, `clear_profile_picture`

        **Response 200:** Array seluruh data pengguna setelah diupdate
      parameters:
      - description: ID pengguna
        in: path
        name: user_id
        required: true
        type: string
      - description: Data yang diperbarui
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_request.UserUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-array_github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
      security:
      - BearerAuth: []
      summary: Update pengguna
      tags:
      - Users
  /users/{user_id}/reset-password:
    post:
      consumes:
      - application/json
      description: |-
        Admin mereset password pengguna tertentu tanpa perlu mengetahui password lama.

        **Path Param:**
        - `user_id` (UUID): ID pengguna yang akan direset passwordnya

        **Request Body:**
        - `new_password` (string, required, min 8 karakter): Password baru untuk pengguna tersebut

        **Akses:** Hanya admin

        **Response 200:** ID pengguna yang passwordnya berhasil direset
      parameters:
      - description: ID pengguna
        in: path
        name: user_id
        required: true
        type: string
      - description: Password baru
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_request.AdminResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common_ActionSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
      security:
      - BearerAuth: []
      summary: Reset password pengguna (Admin)
      tags:
      - Users
  /users/me:
    get:
      description: |-
        Mendapatkan data profil lengkap pengguna yang sedang login.

        **Response 200:** Data pengguna lengkap termasuk profil karyawan (posisi, departemen, gaji pokok, foto profil, dll)

        **Error:** 401 jika token tidak valid atau tidak ada
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
      security:
      - BearerAuth: []
      summary: Profil saya
      tags:
      - Users
    put:
      consumes:
      - application/json
      description: |-
        Memperbarui data profil pengguna yang sedang login. Semua field bersifat opsional (partial update).

        **Field yang bisa diupdate:**
        - `full_name`, `email`, `phone_number`, `address`, `gender`, `birth_date`, `birth_place`
        - `employee_code`, `employment_status`, `position`, `department`, `bank_account_number`
        - `profile_picture_id` (UUID file yang sudah diupload), `clear_profile_picture` (boolean, untuk hapus foto)

        **Response 200:** ID pengguna yang diupdate
      parameters:
      - description: Data profil yang diperbarui
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_request.UserUpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common_ActionSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
      security:
      - BearerAuth: []
      summary: Update profil saya
      tags:
      - Users
  /users/me/password:
    put:
      consumes:
      - application/json
      description: |-
        Mengganti password akun pengguna yang sedang login.

        **Request Body:**
        - `current_password` (string, required): Password saat ini
        - `new_password` (string, required, min 8 karakter): Password baru
        - `confirm_password` (string, required): Konfirmasi password baru, harus sama dengan `new_password`

        **Validasi:**
        - `current_password` harus cocok dengan password yang tersimpan
        - `new_password` dan `confirm_password` harus identik

        **Response 200:** ID pengguna

        **Error:** 400 jika password lama salah atau konfirmasi tidak cocok
      parameters:
      - description: Password lama dan baru
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_request.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common_ActionSuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_afifudin23_absensi-king-royal-api_internal_delivery_http_response_common.Response-any'
      security:
      - BearerAuth: []
      summary: Ganti password
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token JWT dengan format: **Bearer &lt;token&gt;** — contoh:
      `Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...